backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
django-redis==5.4.0
redis==5.0.1
pyecharts==2.0.4 